# ARROW-9928: [C++] Speed up integer parsing slightly