By exiting early out of the parsing routine when the input is exhausted, we can save a little bit a processing time.