@pitrou 2020-09-03
Hi and thank you for posting a PR. I recommend you read a bit about the contribution guidelines here:
https://arrow.apache.org/docs/developers/contributing.html#report-bugs-and-propose-features

Also, can you explain in which situation this marks an improvement? Any benchmark numbers perhaps?