@pitrou 2020-09-07
Integer parsing micro-benchmarks:
* before:
```
IntegerParsing<Int8Type>            2553 ns         2553 ns       826045 items_per_second=391.727M/s
IntegerParsing<Int16Type>           4728 ns         4727 ns       444020 items_per_second=211.536M/s
IntegerParsing<Int32Type>           9060 ns         9059 ns       231401 items_per_second=110.387M/s
IntegerParsing<Int64Type>          12903 ns        12901 ns       162856 items_per_second=77.5155M/s
IntegerParsing<UInt8Type>           3534 ns         3534 ns       587574 items_per_second=282.989M/s
IntegerParsing<UInt16Type>          2135 ns         2135 ns       974843 items_per_second=468.364M/s
IntegerParsing<UInt32Type>          5658 ns         5657 ns       371323 items_per_second=176.765M/s
IntegerParsing<UInt64Type>          7650 ns         7649 ns       274310 items_per_second=130.737M/s
```
* after:
```
IntegerParsing<Int8Type>            2407 ns         2406 ns       879327 items_per_second=415.55M/s
IntegerParsing<Int16Type>           4433 ns         4432 ns       473344 items_per_second=225.632M/s
IntegerParsing<Int32Type>           8648 ns         8647 ns       243097 items_per_second=115.65M/s
IntegerParsing<Int64Type>          12066 ns        12064 ns       174143 items_per_second=82.8881M/s
IntegerParsing<UInt8Type>           3303 ns         3303 ns       637974 items_per_second=302.789M/s
IntegerParsing<UInt16Type>          1996 ns         1996 ns      1052083 items_per_second=501.06M/s
IntegerParsing<UInt32Type>          5301 ns         5300 ns       396027 items_per_second=188.676M/s
IntegerParsing<UInt64Type>          7160 ns         7159 ns       293150 items_per_second=139.681M/s
```