@pitrou 2020-09-07
CSV conversion micro-benchmark:
* before:
```
Int64Conversion     100800 ns       100782 ns        21238 items_per_second=82.6831M/s
```
* after:
```
Int64Conversion      94717 ns        94701 ns        22247 items_per_second=87.9932M/s
```